Add the current theme-plugin.rb plugin (renamed ‘heading-id-filter') to this repository
