+++ wai-website-plugin.gemspec
@@ -2,14 +2,18 @@
 
 Gem::Specification.new do |spec|
   spec.name          = "wai-website-plugin"
-  spec.version       = "0.1.2"
-  spec.authors       = ["Eric Eggert"]
-  spec.email         = ["ee@w3.org"]
+  spec.version       = "0.3"
+  spec.authors       = ["Eric Eggert, Rémi Bétin"]
+  spec.email         = ["wai@w3.org"]
 
-  spec.summary       = ""
-  spec.description   = ""
-  spec.homepage      = "https://www.w3.org/WAI/"
-  spec.license       = "MIT"
+  spec.summary       = "Plugin used for the W3C WAI website"
+  spec.homepage      = "https://github.com/w3c/wai-website-plugin"
+  spec.license       = "Nonstandard"
 
-  spec.files         = ["lib/wai-website-plugin.rb"]
+  spec.files         = ["lib/wai-website-plugin.rb", "lib/wai-website-plugin/heading-id-filter.rb", "lib/wai-website-plugin/links-hooks.rb"]
+
+  spec.required_ruby_version = ">= 3.3.3"
+
+  spec.add_runtime_dependency "jekyll", "~>4.4.1"
+  spec.add_runtime_dependency "nokogiri", '~>1.19.0'
 end
